updated the finance_calculator.py file

+++ python_introduction/finance_calculator.py
@@ -2,8 +2,8 @@
 
 monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your monthly expenses: ")
-monthly_savings = int(monthly_income) - int(monthly_expenses)
-projected_savings = int(monthly_savings * 12 +(monthly_savings * 0.05 * 12))
+monthly_savings = float(monthly_income) - float(monthly_expenses)
+projected_savings = float(monthly_savings * 12 +(monthly_savings * 0.05 * 12))
 
 print("Your Monthly savings are:", monthly_savings)
 print("Your Projected savings for the year are:", projected_savings)
